Move submodule shit to directory
