+++ src/autotrace/.gitattributes
