+++ src/autotrace/autotrace.c
