+++ src/autotrace/autotrace.h
