+++ src/autotrace/bitmap.c
