+++ src/autotrace/bitmap.h
