+++ src/autotrace/color.c
