+++ src/autotrace/color.h
