+++ src/autotrace/curve.c
