+++ src/autotrace/curve.h
