+++ src/autotrace/despeckle.c
