+++ src/autotrace/despeckle.h
