+++ src/autotrace/epsilon-equal.c
