+++ src/autotrace/epsilon-equal.h
