+++ src/autotrace/exception.c
