+++ src/autotrace/exception.h
