+++ src/autotrace/image-header.h
