+++ src/autotrace/image-proc.c
