+++ src/autotrace/image-proc.h
