+++ src/autotrace/input-bmp.c
