+++ src/autotrace/input-bmp.h
