+++ src/autotrace/input-png.c
