+++ src/autotrace/input-png.h
