+++ src/autotrace/input.c
