+++ src/autotrace/logreport.c
