+++ src/autotrace/logreport.h
