+++ src/autotrace/module.c
