+++ src/autotrace/private.h
