+++ src/autotrace/pxl-outline.c
