+++ src/autotrace/pxl-outline.h
