+++ src/autotrace/spline.c
