+++ src/autotrace/spline.h
