+++ src/autotrace/thin-image.c
