+++ src/autotrace/thin-image.h
