+++ src/autotrace/types.h
