+++ src/autotrace/vector.c
