+++ src/autotrace/vector.h
